+++ fundamentals/today-i-learned/src/api/remote/reactions.ts
@@ -1,7 +1,7 @@
 // 리액션 관련 원격 API 함수
 
-import { graphqlRequest } from "../client";
-import { ADD_REACTION_MUTATION, REMOVE_REACTION_MUTATION } from "../graphql/reactions";
+import { graphqlRequest } from "@/api/client";
+import { ADD_REACTION_MUTATION, REMOVE_REACTION_MUTATION } from "@/api/graphql/reactions";
 
 export interface AddReactionParams {
   discussionId: string;
