+++ fundamentals/today-i-learned/src/components/features/challenge/MonthlyChallenge.tsx
