+++ fundamentals/today-i-learned/src/components/features/sprint/SprintDayItem.tsx
