+++ fundamentals/today-i-learned/src/libs/sprintCalculator.ts
